chore: moving the dictionary to its own file
